+++ Source/PropertyTools.Wpf/DataGrid/Operators/ListOperator.cs
@@ -216,7 +216,7 @@ namespace PropertyTools.Wpf
         public override string GetBindingPath(CellRef cell)
         {
             var pd = this.GetPropertyDefinition(cell);
-            if (pd?.PropertyName != null)
+            if (!string.IsNullOrEmpty(pd?.PropertyName))
             {
                 return pd.PropertyName;
             }